+++ agents-manage-ui/src/components/form/expandable-field.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Maximize } from 'lucide-react';
-import type { ComponentProps, ReactNode } from 'react';
+import { type ComponentProps, type ReactNode, useCallback } from 'react';
 import { Button } from '@/components/ui/button';
 import {
   Dialog,
@@ -11,12 +11,13 @@ import {
   DialogTrigger,
 } from '@/components/ui/dialog';
 import { Label } from '@/components/ui/label';
+import { useMonacoStore } from '@/features/agent/state/use-monaco-store';
 import { cn } from '@/lib/utils';
 
 type DialogProps = Required<ComponentProps<typeof Dialog>>;
 
 interface ExpandableFieldProps {
-  name: string;
+  uri: string;
   label: string;
   className?: string;
   children: ReactNode;
@@ -24,21 +25,36 @@ interface ExpandableFieldProps {
   isRequired?: boolean;
   open: DialogProps['open'];
   onOpenChange: DialogProps['onOpenChange'];
+  hasError?: boolean;
+  id: string;
 }
 
 export function ExpandableField({
-  name,
+  id,
+  uri,
   label,
   children,
   actions,
   isRequired = false,
   open,
   onOpenChange,
+  hasError,
 }: ExpandableFieldProps) {
+  const monaco = useMonacoStore((state) => state.monaco);
+
+  const handleClick = useCallback(() => {
+    if (!monaco) {
+      return;
+    }
+    const model = monaco.editor.getModel(monaco.Uri.parse(uri));
+    const [editor] = monaco.editor.getEditors().filter((editor) => editor.getModel() === model);
+    editor?.focus();
+  }, [monaco, uri]);
+
   const content = (
     <>
       <div className="flex items-center justify-between">
-        <Label className="gap-1" htmlFor={name}>
+        <Label id={id} className={cn(hasError && 'text-red-600', 'gap-1')} onClick={handleClick}>
           {label}
           {isRequired && <span className="text-red-500">*</span>}
         </Label>
@@ -54,7 +70,7 @@ export function ExpandableField({
           )}
         </div>
       </div>
-      <div className={cn('relative', open && 'grow')}>{children}</div>
+      <div className={cn('relative space-y-2', open && 'grow')}>{children}</div>
     </>
   );